modals.py is singled out.
